[01] Python/Code with Hasan/requirements.txt
streamlit
cohere
openai
nlpcloud
tiktoken
youtube-transcript-api
sqlalchemy
chardet
google-cloud-bigquery
pytube
matplotlib
seaborn
plotly
yfinance
mplfinance
numpy
pandas
pyodbc
openpyxl
langchain
langchain_community
langchain_openai
langchain_core
langgraph
langserve
fastapi
uvicorn
Chromadb
faiss-cpu
azure-storage-blob
langchain-text-splitters